requirements.txt
pip install torch==2.0.1 torchvision==0.15.2 torchaudio==2.0.2 --index-url https://download.pytorch.org/whl/cu118
timm==0.9.2
yacs==0.1.8
opencv-python==4.8.1.78
termcolor==2.3.0
numpy==1.24.3
matplotlib==3.7.2
tqdm==4.66.1
tensorboard==2.13.0
torchreid == 0.2.5
PySocks-1.7.1 beautifulsoup4-4.14.3 gdown-6.0.0 soupsieve-2.8.3